Feature: Implement data-related logic for RangeData and update range factory methods for consistency
- Introduced RangeData abstraction to couple logical ranges with data sequences and domains.
- Updated RangeFactory methods to remove type parameters for cleaner usage.
- Added TryCreate method for Range with validation and error messaging.
- Enhanced Distance method in domain interfaces to ensure accurate step calculations.
- Revised tests to reflect changes in range factory methods and ensure correctness.
- Updated project metadata and CI/CD configurations for new data module.

+++ src/Domain/Intervals.NET.Domain.Default/Calendar/StandardDateOnlyBusinessDaysVariableStepDomain.cs
@@ -163,17 +163,17 @@ public readonly struct StandardDateOnlyBusinessDaysVariableStepDomain : IVariabl
     /// <para>⚠️ <strong>Performance:</strong> O(N) - Iterates through each day in the range.</para>
     /// </remarks>
     [Pure]
-    public double Distance(DateOnly start, DateOnly end)
+    public long Distance(DateOnly start, DateOnly end)
     {
         var current = Floor(start);
         var target = Floor(end);
 
         if (current == target)
         {
-            return 0.0; // Same date = 0 steps needed
+            return 0; // Same date = 0 steps needed
         }
 
-        double count = 0;
+        long count = 0;
 
         if (current < target)
         {
